+++ tests/testthat/test-geom_pointpath.R
@@ -13,13 +13,30 @@ test_that("geom_pointpath plots can be build", {
   expect_is(gt, "gtable")
   grob <- layer_grob(g)[[1]]$children
   expect_is(grob[[2]], "points")
-  expect_is(grob[[1]], "pointpath")
+  expect_is(grob[[1]], "gapsegments")
 })
 
-test_that("geom_pointpath makeContent works", {
+test_that("geom_pointpath makeContext works", {
   g <- base + geom_pointpath()
   grob <- layer_grob(g)[[1]]$children[[1]]
-  expect_is(grob, "pointpath")
-  out <- grid::makeContent(grob)
+  expect_is(grob, "gapsegments")
+  out <- grid::makeContext(grob)
+  expect_is(out, "segments")
+})
+
+test_that("geom_pointpath plots can be build in polar coordinates", {
+  g <- base + geom_pointpath() + coord_polar()
+  gt <- ggplotGrob(g)
+  expect_is(gt, "gtable")
+  grob <- layer_grob(g)[[1]]$children
+  expect_is(grob[[2]], "points")
+  expect_is(grob[[1]], "gapsegmentschain")
+})
+
+test_that("geom_pointpath makeContext works with polar coordinates", {
+  g <- base + geom_pointpath() + coord_polar()
+  grob <- layer_grob(g)[[1]]$children[[1]]
+  expect_is(grob, "gapsegmentschain")
+  out <- grid::makeContext(grob)
   expect_is(out, "segments")
 })